+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/runtime/PMDRuntimeConstants.java
@@ -31,6 +31,11 @@ public class PMDRuntimeConstants {
     public static final String[] ALL_MARKER_TYPES = new String[] { PMD_MARKER, PMD_DFA_MARKER, PMD_TASKMARKER,
         PMD_MARKER_1, PMD_MARKER_2, PMD_MARKER_3, PMD_MARKER_4, PMD_MARKER_5 };
 
+    /**
+     * @deprecated not useful, as the property cannot occur on rules anyway. The default was used always,
+     *             which was 1000 violations.
+     */
+    @Deprecated
     public static final IntegerProperty MAX_VIOLATIONS_DESCRIPTOR = new IntegerProperty("maxviolations",
             "Max allowable violations", 1, Integer.MAX_VALUE - 1, 1000, 0f);